Moved tests around -- for some reason, the Aruco method call is 10x faster than our implementation now -- maybe because I built with CUDA_FAST_MATH??
