+++ ImP/imageProcessing/parallel/tests/testMarkerDetectPar.py
@@ -63,39 +63,6 @@ class TestMarkerDetectPar(unittest.TestCase):
         print(otsu_rv)
         np.testing.assert_allclose(no_otsu, otsu)
 
-    def test_cuda_warp_perspective_equals_warp_perspective(self):
-        # self.fail()
-        candidates, _ = aruco._detectCandidates(self.gray_marker_0, aruco.DetectorParameters_create())
-        corners = candidates[9]
-        src = self.gray_marker_0
-        M = np.array([[ 7.03145227e-03,  5.50015822e-02, -5.41421824e+00],
-                      [-6.65280496e-02, -4.24647125e-04,  2.78278338e+01],
-                      [ 6.38002118e-04, -2.75228447e-04,  1.00000000e+00]], dtype=np.float32)
-        result_img_size = 24
-        result_img_corners = np.array([[ 0.,  0.],
-                                       [23.,  0.],
-                                       [23., 23.],
-                                       [ 0., 23.]])
-        actual_dst = cv2.warpPerspective(src, M, (result_img_size, result_img_size), flags=cv2.INTER_NEAREST)
-        test_dst = GpuWrapper.cudaWarpPerspectiveWrapper(src.astype(dtype=np.uint8),
-                                                         M.astype(dtype=np.float32),
-                                                         (result_img_size, result_img_size),
-                                                         _flags=cv2.INTER_NEAREST)
-        test_host_dst = GpuWrapper.warpPerspectiveWrapper(src.astype(dtype=np.uint8),
-                                                          M.astype(dtype=np.float32),
-                                                          (result_img_size, result_img_size),
-                                                          _flags=cv2.INTER_NEAREST)
-        np.testing.assert_array_equal(actual_dst, test_host_dst)
-        np.testing.assert_array_equal(actual_dst, test_dst)
-
-
-    def test_echo_from_cython(self):
-        M = np.array([[7.03145227e-03, 5.50015822e-02, -5.41421824e+00],
-                      [-6.65280496e-02, -4.24647125e-04, 2.78278338e+01],
-                      [6.38002118e-04, -2.75228447e-04, 1.00000000e+00]], dtype=np.float32)
-        test_M = GpuWrapper.echoPyObject(M)
-        np.testing.assert_allclose(M, test_M)
-
     # PUBLIC FUNCTIONS
 
     def test_detector_parameters(self):
